refactor(tests): split ipfs-request.test.js
